+++ src/resolvers/VideoAddSemitagEvent/index.ts
@@ -1,9 +1,21 @@
+import z from "zod";
+
 import { Resolvers } from "../graphql.js";
 import { ResolverDeps } from "../index.js";
 import { SemitagModel } from "../Semitag/model.js";
 
+const schemaPayload = z.object({
+  /**
+   * SemitagのID
+   */
+  id: z.string(),
+});
+export type VideoAddSemitagEventPayload = z.infer<typeof schemaPayload>;
+
 export const resolveVideoAddSemitagEvent = ({ prisma }: Pick<ResolverDeps, "prisma">) =>
   ({
     semitag: ({ payload }) =>
-      prisma.semitag.findUniqueOrThrow({ where: { id: payload.id } }).then((v) => new SemitagModel(v)),
+      prisma.semitag
+        .findUniqueOrThrow({ where: { id: schemaPayload.parse(payload).id } })
+        .then((v) => new SemitagModel(v)),
   } satisfies Resolvers["VideoAddSemitagEvent"]);
