+++ src/resolvers/VideoRemoveThumbnailEvent/index.ts
@@ -1,5 +1,10 @@
+import z from "zod";
+
 import { Resolvers } from "../graphql.js";
 import { ResolverDeps } from "../index.js";
+
+const schemaPayload = z.object({ id: z.string() });
+export type VideoRemoveThumbnailEventPayload = z.infer<typeof schemaPayload>;
 
 export const resolveVideoRemoveThumbnailEvent = ({ prisma }: Pick<ResolverDeps, "prisma">) =>
   ({} satisfies Resolvers["VideoRemoveThumbnailEvent"]);
